refactor(agent): normalize complete publish identity plans
