+++ packages/agent/src/publish-author-selection.ts
@@ -1,4 +1,5 @@
 import { PUBLISH_AUTHOR_SELECTION_CONFLICT_CODE } from '@origintrail-official/dkg-core';
+import type { ResidentAssertionAuthorSelection } from './finalized-assertion-author.js';
 
 /** Identity choice for publishing an already-finalized assertion. */
 export type PublishAuthorSelection =
@@ -28,22 +29,80 @@ export type PublishAuthorSelectionOptions =
     selectedAuthorAgentAddress?: string;
   };
 
-type NormalizedPublishAuthorSelection =
-  | PublishAuthorSelection
-  | {
-    readonly mode: 'residentAuthor';
-    readonly selectedAuthorAgentAddress: unknown;
-    readonly callerAgentAddress?: string;
-    readonly agentAddress?: never;
-  }
-  | { readonly mode: 'default' };
+/** All identity decisions consumed by author lookup and durable enqueue. */
+export interface PublishIdentityPlan {
+  readonly author:
+    | { readonly mode: 'author'; readonly agentAddress: string }
+    | {
+        readonly mode: 'resolve';
+        readonly callerHint: string;
+        readonly residentSelection: ResidentAssertionAuthorSelection | undefined;
+      };
+  readonly enqueueCaller: string | undefined;
+}
 
 function conflict(message: string): never {
   throw Object.assign(new Error(message), { code: PUBLISH_AUTHOR_SELECTION_CONFLICT_CODE });
 }
 
-/** Snapshot the selection before author lookup; reject contradictory untyped input. */
-export function readPublishAuthorSelection(options?: PublishAuthorSelectionOptions): NormalizedPublishAuthorSelection {
+/** Snapshot untyped API input without forwarding arbitrary values into lookup. */
+export function readResidentAuthorSelection(value: unknown): ResidentAssertionAuthorSelection | undefined {
+  if (value === undefined) return undefined;
+  return typeof value === 'string'
+    ? Object.freeze({ kind: 'address', agentAddress: value })
+    : Object.freeze({ kind: 'malformed', displayValue: String(value) });
+}
+
+function authorPlan(agentAddress: string, enqueueCaller: string | undefined): PublishIdentityPlan {
+  return Object.freeze({
+    author: Object.freeze({ mode: 'author', agentAddress }),
+    enqueueCaller: enqueueCaller || undefined,
+  });
+}
+
+function lookupPlan(
+  callerHint: string,
+  enqueueCaller: string | undefined,
+  residentSelection?: ResidentAssertionAuthorSelection,
+): PublishIdentityPlan {
+  return Object.freeze({
+    author: Object.freeze({ mode: 'resolve', callerHint, residentSelection }),
+    enqueueCaller: enqueueCaller || undefined,
+  });
+}
+
+/** Adapt released flat-field quirks once, before any asynchronous author lookup. */
+function readLegacyPublishIdentityPlan(
+  agentAddress: unknown,
+  callerAgentAddress: unknown,
+  selectedAuthorAgentAddress: unknown,
+  defaultCallerHint: string,
+): PublishIdentityPlan {
+  if ((agentAddress !== undefined && typeof agentAddress !== 'string')
+    || (callerAgentAddress !== undefined && typeof callerAgentAddress !== 'string')) {
+    return conflict('Invalid or conflicting VM publish author selection fields');
+  }
+  // An empty authoritative selector is ignored. An explicitly empty caller
+  // beside a nonempty author override suppresses the enqueue caller stamp.
+  const enqueueCaller = callerAgentAddress ?? agentAddress;
+  if (agentAddress) {
+    if (callerAgentAddress || selectedAuthorAgentAddress !== undefined) {
+      return conflict('Invalid or conflicting VM publish author selection fields');
+    }
+    return authorPlan(agentAddress, enqueueCaller);
+  }
+  return lookupPlan(
+    callerAgentAddress ?? defaultCallerHint,
+    enqueueCaller,
+    readResidentAuthorSelection(selectedAuthorAgentAddress),
+  );
+}
+
+/** Normalize either public syntax into one immutable, behavior-complete plan. */
+export function readPublishIdentityPlan(
+  options: PublishAuthorSelectionOptions | undefined,
+  defaultCallerHint: string,
+): PublishIdentityPlan {
   const rawOptions = options as Record<string, unknown> | undefined;
   const selection = rawOptions?.authorSelection;
   const agentAddress = rawOptions?.agentAddress;
@@ -55,30 +114,11 @@ export function readPublishAuthorSelection(options?: PublishAuthorSelectionOptio
   if (selection !== undefined && hasFlatSelection) {
     return conflict('VM publish identity fields must use either authorSelection or the compatible flat form');
   }
-  if (selection === undefined && hasFlatSelection) {
-    if ((agentAddress !== undefined && typeof agentAddress !== 'string')
-      || (callerAgentAddress !== undefined && typeof callerAgentAddress !== 'string')) {
-      return conflict('Invalid or conflicting VM publish author selection fields');
-    }
-    // The released flat contract ignores an empty authoritative selector and
-    // permits an empty caller beside a nonempty author override.
-    if (agentAddress) {
-      if (callerAgentAddress || selectedAuthorAgentAddress !== undefined) {
-        return conflict('Invalid or conflicting VM publish author selection fields');
-      }
-      return Object.freeze({ mode: 'author', agentAddress });
-    }
-    if (selectedAuthorAgentAddress !== undefined) {
-      return Object.freeze({
-        mode: 'residentAuthor',
-        selectedAuthorAgentAddress,
-        ...(callerAgentAddress === undefined ? {} : { callerAgentAddress }),
-      });
-    }
-    if (callerAgentAddress === undefined) return Object.freeze({ mode: 'default' });
-    return Object.freeze({ mode: 'callerHint', callerAgentAddress });
+  if (selection === undefined) {
+    return readLegacyPublishIdentityPlan(
+      agentAddress, callerAgentAddress, selectedAuthorAgentAddress, defaultCallerHint,
+    );
   }
-  if (selection === undefined) return Object.freeze({ mode: 'default' });
   if (selection === null || typeof selection !== 'object') return conflict('Invalid VM publish authorSelection');
   const {
     mode,
@@ -90,27 +130,19 @@ export function readPublishAuthorSelection(options?: PublishAuthorSelectionOptio
     case 'author':
       if (typeof nestedAgentAddress !== 'string' || nestedAgentAddress.length === 0
         || nestedCallerAgentAddress !== undefined || nestedSelectedAuthorAgentAddress !== undefined) break;
-      return Object.freeze({ mode, agentAddress: nestedAgentAddress });
+      return authorPlan(nestedAgentAddress, nestedAgentAddress);
     case 'callerHint':
       if (typeof nestedCallerAgentAddress !== 'string'
         || nestedAgentAddress !== undefined || nestedSelectedAuthorAgentAddress !== undefined) break;
-      return Object.freeze({ mode, callerAgentAddress: nestedCallerAgentAddress });
+      return lookupPlan(nestedCallerAgentAddress, nestedCallerAgentAddress);
     case 'residentAuthor':
       if (nestedAgentAddress !== undefined || nestedSelectedAuthorAgentAddress === undefined
         || (nestedCallerAgentAddress !== undefined && typeof nestedCallerAgentAddress !== 'string')) break;
-      // Resident-selector validation stays with the canonical assertion-author
-      // resolver, including its existing ASSERTION_AUTHOR_NOT_RESIDENT errors.
-      return Object.freeze({
-        mode,
-        selectedAuthorAgentAddress: nestedSelectedAuthorAgentAddress,
-        ...(nestedCallerAgentAddress === undefined ? {} : { callerAgentAddress: nestedCallerAgentAddress }),
-      });
+      return lookupPlan(
+        nestedCallerAgentAddress ?? defaultCallerHint,
+        nestedCallerAgentAddress,
+        readResidentAuthorSelection(nestedSelectedAuthorAgentAddress),
+      );
   }
   return conflict('Invalid or conflicting VM publish authorSelection fields');
-}
-
-/** Preserve the enqueuing caller separately from the resolved member author. */
-export function publishAuthorCallerIdentity(selection: NormalizedPublishAuthorSelection): string | undefined {
-  if (selection.mode === 'author') return selection.agentAddress;
-  return selection.mode === 'default' ? undefined : selection.callerAgentAddress;
 }
