+++ packages/agent/test/publish-foreign-author-resolution.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, expectTypeOf, vi } from 'vitest';
+import { describe, it, expect } from 'vitest';
 import {
   buildAssertionSealQuads,
   contextGraphAssertionUri,
@@ -7,14 +7,8 @@ import {
   ASSERTION_SEAL_PREDICATES,
   GRAPH_KA_CONTENT_SCOPE_VERSION,
 } from '@origintrail-official/dkg-core';
-import { GraphManager, OxigraphStore, type Quad } from '@origintrail-official/dkg-storage';
-import {
-  computeFlatKCRootV10,
-  storeKnowledgeAssetOperationPublicQuads,
-  storeKnowledgeAssetWorkspaceHead,
-} from '@origintrail-official/dkg-publisher';
-import { DKGAgent } from '../src/dkg-agent.js';
-import { readPublishAuthorSelection, type PublishAuthorSelectionOptions } from '../src/publish-author-selection.js';
+import { OxigraphStore, type Quad } from '@origintrail-official/dkg-storage';
+import { CG, MEMBER, CURATOR, OTHER, NAME, KA_UAL, RESERVED_KA_ID, PUBLIC_QUAD, MERKLE, sealAt, sealFor, stubAgent } from './_helpers/finalized-author.js';
 
 /**
  * GH#1778 — a curator publishes a rootless named KA authored by a MEMBER and
@@ -24,169 +18,6 @@ import { readPublishAuthorSelection, type PublishAuthorSelectionOptions } from '
  * the caller. These tests pin `resolveAssertionAuthor` and the end-to-end
  * auto-resolution in `publishFromFinalizedAssertion`.
  */
-
-const CG = 'construction';
-const MEMBER = '0xA32f1cc125401B55911678847426759094055B2d';
-const CURATOR = `0x${'11'.repeat(20)}`;
-const OTHER = `0x${'22'.repeat(20)}`;
-const NAME = 'justTriplets';
-const KA_UAL = `did:dkg:hardhat:31337/${MEMBER}/7`;
-const RESERVED_KA_ID = (BigInt(MEMBER) << 96n) | 7n;
-const PUBLIC_QUAD: Quad = {
-  subject: 'urn:justTriplets:subject1',
-  predicate: 'urn:justTriplets:predicate1',
-  object: '"value1"',
-  graph: '',
-};
-const MERKLE = computeFlatKCRootV10([PUBLIC_QUAD], []);
-
-it.each([
-  { authorSelection: null },
-  { authorSelection: { mode: 'unknown' } },
-  { authorSelection: { mode: 'default' } },
-  { authorSelection: { mode: 'callerHint' } },
-  { authorSelection: { mode: 'residentAuthor', callerAgentAddress: CURATOR } },
-  { authorSelection: { mode: 'callerHint', callerAgentAddress: CURATOR }, callerAgentAddress: CURATOR },
-  { callerAgentAddress: 7 },
-  { callerAgentAddress: 7, selectedAuthorAgentAddress: MEMBER },
-  { subGraphName: 'research', agentAddress: OTHER, callerAgentAddress: CURATOR },
-])('rejects malformed or contradictory untyped author selection before looking up an author: %j', async options => {
-  const store = new OxigraphStore();
-  const query = vi.spyOn(store, 'query');
-  const agent = stubAgent(store, CURATOR);
-  await expect(agent.resolveFinalizedAssertionPublishAuthor(CG, NAME, options as never))
-    .rejects.toMatchObject({ code: 'PUBLISH_AUTHOR_SELECTION_CONFLICT' });
-  expect(query).not.toHaveBeenCalled();
-});
-
-it.each([null, 42, {}])('keeps a malformed resident selector unvalidated in the internal type: %j', selectedAuthorAgentAddress => {
-  const selection = readPublishAuthorSelection({ authorSelection: { mode: 'residentAuthor', selectedAuthorAgentAddress } } as never);
-  expect(selection.mode).toBe('residentAuthor');
-  if (selection.mode === 'residentAuthor') {
-    expectTypeOf(selection.selectedAuthorAgentAddress).toEqualTypeOf<unknown>();
-    expect(selection.selectedAuthorAgentAddress).toBe(selectedAuthorAgentAddress);
-  }
-});
-
-it('treats undefined legacy optional fields as the default selection', async () => {
-  const store = new OxigraphStore();
-  await store.insert(sealFor(MEMBER));
-  const agent = stubAgent(store, CURATOR);
-  expect(await agent.resolveFinalizedAssertionPublishAuthor(CG, NAME, {
-    agentAddress: undefined, callerAgentAddress: undefined, selectedAuthorAgentAddress: undefined,
-  })).toBe(MEMBER);
-});
-
-function sealAt(cg: string, author: string, name = NAME, subGraphName?: string): Quad[] {
-  return buildAssertionSealQuads({
-    assertionUri: contextGraphAssertionUri(cg, author, name, subGraphName),
-    metaGraph: contextGraphMetaUri(cg),
-    merkleRoot: MERKLE,
-    authorAddress: author,
-    authorAttestationR: new Uint8Array(32).fill(1),
-    authorAttestationVS: new Uint8Array(32).fill(2),
-    authorSchemeVersion: 1,
-    chainId: 31337n,
-    kav10Address: '0x1234567890123456789012345678901234567890',
-    reservedKaId: (BigInt(author) << 96n) | 7n,
-    finalizedAtIso: '2026-01-01T00:00:00.000Z',
-    contentScopeVersion: GRAPH_KA_CONTENT_SCOPE_VERSION,
-    kaUal: `did:dkg:hardhat:31337/${author}/7`,
-    assertionVersion: 1,
-    publicTripleCount: 1,
-    privateTripleCount: 0,
-  }) as Quad[];
-}
-
-function sealFor(author: string, name = NAME): Quad[] {
-  return sealAt(CG, author, name);
-}
-
-function makeLog() {
-  return { debug() {}, info() {}, warn() {}, error() {} };
-}
-
-function stubAgent(store: OxigraphStore, defaultAgentAddress: string) {
-  const agent = Object.create(DKGAgent.prototype) as any;
-  agent.store = store;
-  agent.log = makeLog();
-  agent.defaultAgentAddress = defaultAgentAddress;
-  Object.defineProperty(agent, 'peerId', {
-    value: '12D3KooWQz2bQbQueABKRSjV9koF8VYsXk5TdCsUmPf5zAEZg3q6',
-    configurable: true,
-  });
-  return agent;
-}
-
-/** Real seals and operation heads; only the chain submission and history facade are stubbed. */
-async function publishableAgent() {
-  const store = new OxigraphStore();
-  const graphManager = new GraphManager(store);
-  for (const author of [MEMBER, CURATOR]) {
-    const kaUal = `did:dkg:hardhat:31337/${author}/7`;
-    const scope = { store, graphManager, contextGraphId: CG, kaUal, assertionVersion: 1, shareOperationId: `share-${author}` };
-    await store.insert([...sealFor(author), { ...PUBLIC_QUAD, graph: `${contextGraphSharedMemoryUri(CG)}/${author.toLowerCase()}/7` }]);
-    await storeKnowledgeAssetWorkspaceHead(scope);
-    await storeKnowledgeAssetOperationPublicQuads({ ...scope, quads: [PUBLIC_QUAD], accessPolicy: 'public' });
-  }
-  const agent = stubAgent(store, CURATOR);
-  agent.chain = {};
-  agent.publisher = { hasSwmShareComplete: async () => true, clearSwmShareComplete: async () => {}, clearRemainingSharedMemory: async () => {} };
-  Object.defineProperty(agent, 'assertion', {
-    value: { history: async (_cg: string, _name: string, options: { agentAddress: string }) => ({ events: [], currentShareOperationId: `share-${options.agentAddress}` }) },
-  });
-  agent.publishFromSharedMemory = vi.fn(async () => ({ kaId: RESERVED_KA_ID, ual: KA_UAL, merkleRoot: MERKLE, kaManifest: [], status: 'confirmed', publicQuads: [] }));
-  return { agent, store };
-}
-
-it.each<{ options: PublishAuthorSelectionOptions; author: string; caller?: string }>([
-  { options: { agentAddress: MEMBER }, author: MEMBER, caller: MEMBER },
-  { options: { callerAgentAddress: MEMBER }, author: MEMBER, caller: MEMBER },
-  { options: { selectedAuthorAgentAddress: MEMBER, callerAgentAddress: CURATOR }, author: MEMBER, caller: CURATOR },
-  { options: { selectedAuthorAgentAddress: MEMBER }, author: MEMBER },
-  { options: { agentAddress: '', callerAgentAddress: MEMBER }, author: MEMBER, caller: MEMBER },
-  { options: { agentAddress: MEMBER, callerAgentAddress: '' }, author: MEMBER },
-  { options: { agentAddress: '' }, author: CURATOR },
-  { options: {}, author: CURATOR },
-])('preserves released flat-option behavior across all three public methods: $options', async ({ options, author, caller }) => {
-  const { agent } = await publishableAgent();
-  expect(await agent.resolveFinalizedAssertionPublishAuthor(CG, NAME, options)).toBe(author);
-  const intent = await agent.resolveFinalizedAssertionVmPublishIntent(CG, NAME, options);
-  expect(intent.agentAddress).toBe(author);
-  expect(intent.callerAgentAddress).toBe(caller);
-  const published = await agent.publishFromFinalizedAssertion(CG, NAME, options);
-  expect(published.seal.authorAddress).toBe(author);
-  expect(agent.publishFromSharedMemory).toHaveBeenCalledWith(CG, expect.anything(), expect.objectContaining({ precomputedAttestation: expect.objectContaining({ authorAddress: author }) }));
-});
-
-it.each(['callerHint', 'residentAuthor', 'legacy'] as const)('snapshots %s author and caller across an awaited lookup', async mode => {
-  const { agent, store } = await publishableAgent();
-  const authorSelection = { mode: 'residentAuthor' as const, selectedAuthorAgentAddress: MEMBER, callerAgentAddress: CURATOR };
-  const callerSelection = { mode: 'callerHint' as const, callerAgentAddress: MEMBER };
-  const legacy = { callerAgentAddress: MEMBER };
-  const options = mode === 'residentAuthor' ? { authorSelection } : mode === 'callerHint' ? { authorSelection: callerSelection } : legacy;
-  let release!: () => void;
-  let entered!: () => void;
-  const blocked = new Promise<void>(resolve => { release = resolve; });
-  const lookupStarted = new Promise<void>(resolve => { entered = resolve; });
-  const query = store.query.bind(store);
-  vi.spyOn(store, 'query').mockImplementationOnce(async (...args) => {
-    entered();
-    await blocked;
-    return query(...args);
-  });
-  const pending = agent.resolveFinalizedAssertionVmPublishIntent(CG, NAME, options);
-  await lookupStarted;
-  authorSelection.selectedAuthorAgentAddress = CURATOR;
-  authorSelection.callerAgentAddress = OTHER;
-  callerSelection.callerAgentAddress = CURATOR;
-  legacy.callerAgentAddress = CURATOR;
-  release();
-  const intent = await pending;
-  expect(intent.agentAddress).toBe(MEMBER);
-  expect(intent.seal.authorAddress.toLowerCase()).toBe(MEMBER.toLowerCase());
-  expect(intent.callerAgentAddress).toBe(mode === 'residentAuthor' ? CURATOR : MEMBER);
-});
 
 describe('GH#1778 resolveAssertionAuthor', () => {
   it('resolves the sole (member) author when the caller (curator) is not the author', async () => {
@@ -815,76 +646,6 @@ describe('GH#1786 selectedAuthorAgentAddress (resident-candidate selection)', ()
 });
 
 describe('GH#1778 resolveFinalizedAssertionVmPublishIntent (async) auto-resolves the member author', () => {
-  it('snapshots author selection before an asynchronous author lookup', async () => {
-    const store = new OxigraphStore();
-    await store.insert(sealFor(MEMBER));
-    const graphManager = new GraphManager(store);
-    const shareOperationId = 'selection-snapshot';
-    await storeKnowledgeAssetOperationPublicQuads({
-      store,
-      graphManager,
-      contextGraphId: CG,
-      shareOperationId,
-      kaUal: KA_UAL,
-      assertionVersion: 1,
-      quads: [PUBLIC_QUAD],
-      privateTripleCount: 0,
-      publisherPeerId: 'publisher-peer',
-      accessPolicy: 'ownerOnly',
-      agentAddress: MEMBER,
-    });
-    await storeKnowledgeAssetWorkspaceHead({
-      store,
-      graphManager,
-      contextGraphId: CG,
-      shareOperationId,
-      kaUal: KA_UAL,
-      assertionVersion: 1,
-    });
-
-    const agent = stubAgent(store, CURATOR);
-    agent.publisher = { hasSwmShareComplete: async () => true };
-    agent.getCustodialAgentPrivateKey = () => undefined;
-    Object.defineProperty(agent, 'assertion', {
-      value: {
-        history: async () => ({ events: [], currentShareOperationId: shareOperationId }),
-      },
-      configurable: true,
-    });
-
-    let beginLookup!: () => void;
-    const lookupStarted = new Promise<void>((resolve) => { beginLookup = resolve; });
-    let finishLookup!: () => void;
-    const lookupReleased = new Promise<void>((resolve) => { finishLookup = resolve; });
-    const query = store.query.bind(store);
-    let deferred = false;
-    vi.spyOn(store, 'query').mockImplementation((async (sparql: string, ...args: unknown[]) => {
-      if (!deferred && sparql.includes('SELECT DISTINCT ?s')) {
-        deferred = true;
-        beginLookup();
-        await lookupReleased;
-      }
-      return query(sparql, ...args as never[]);
-    }) as typeof store.query);
-
-    const options = {
-      authorSelection: {
-        mode: 'residentAuthor' as const,
-        callerAgentAddress: CURATOR,
-        selectedAuthorAgentAddress: MEMBER,
-      },
-    };
-    const pending = agent.resolveFinalizedAssertionVmPublishIntent(CG, NAME, options);
-    await lookupStarted;
-    options.authorSelection.callerAgentAddress = OTHER;
-    options.authorSelection.selectedAuthorAgentAddress = OTHER;
-    finishLookup();
-
-    const intent = await pending;
-    expect(intent.agentAddress).toBe(MEMBER);
-    expect(intent.callerAgentAddress).toBe(CURATOR);
-  });
-
   it('resolves the member author from _meta when the caller (curator) is not the author', async () => {
     const store = new OxigraphStore();
     await store.insert(sealFor(MEMBER));
